+++ src/app/pages/staff/dashboard/dashboard.component.ts
@@ -1,46 +1,37 @@
-import {Component} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {ApexChart, ApexStroke, ApexXAxis} from "ng-apexcharts";
+import {IStaffDashboardStats} from "@lib/interfaces/istaff-dashboard-stats";
+import {StaffService} from "@lib/services/staff/staff.service";
+import {Notify} from "notiflix/build/notiflix-notify-aio";
+import {StaffDashboardStats} from "@lib/classes/StaffDashboardStats";
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.css']
 })
-export class DashboardComponent {
+export class DashboardComponent implements OnInit {
+  pageData: IStaffDashboardStats = new StaffDashboardStats();
   chart: ApexChart = {
     height: 390,
     type: "area"
   };
   xaxis: ApexXAxis = {
-    type: "datetime",
+    type: "category",
     categories: [
-      "2018-09-19T00:00:00.000Z",
-      "2018-09-19T01:30:00.000Z",
-      "2018-09-19T02:30:00.000Z",
-      "2018-09-19T03:30:00.000Z",
-      "2018-09-19T04:30:00.000Z",
-      "2018-09-19T05:30:00.000Z",
-      "2018-09-19T06:30:00.000Z"
+      "Sun",
+      "Mon",
+      "Tue",
+      "Wed",
+      "Thur",
+      "Fri",
+      "Sat"
     ]
   };
   stroke: ApexStroke = {
     curve: "smooth"
   };
   chartOptions = {
-    series: [
-      {
-        name: "pending",
-        data: [31, 40, 28, 51, 42, 109, 100]
-      },
-      {
-        name: "approved",
-        data: [11, 32, 45, 32, 34, 52, 41]
-      },
-      {
-        name: "declined",
-        data: [21, 15, 0, 50, 12, 30, 80]
-      }
-    ],
     dataLabels: {
       enabled: false
     },
@@ -51,19 +42,18 @@ export class DashboardComponent {
     }
   };
 
-  public generateData(baseval: any, count: number, yrange: any) {
-    let i = 0;
-    const series = [];
-    while (i < count) {
-      const x = Math.floor(Math.random() * (750 - 1 + 1)) + 1;
-      const y =
-        Math.floor(Math.random() * (yrange.max - yrange.min + 1)) + yrange.min;
-      const z = Math.floor(Math.random() * (75 - 15 + 1)) + 15;
+  constructor(private readonly staffService: StaffService) {
+  }
 
-      series.push([x, y, z]);
-      baseval += 86400000;
-      i++;
-    }
-    return series;
+  ngOnInit() {
+    this.staffService.retrieveDashboardStats().subscribe((res) => {
+      this.pageData = res;
+    }, () => {
+      Notify.failure('An error occurred')
+    })
+  }
+
+  calcPercent(value: number) {
+    return Number(value / this.pageData.submissions) * 100
   }
 }